Code refactoring. Placed blocks now land at a position.

+++ core/src/com/syurba/combogame/FallingBlock.java
@@ -3,7 +3,7 @@ package com.syurba.combogame;
 import com.badlogic.gdx.math.Rectangle;
 
 public class FallingBlock extends Rectangle {
-    public FallingBlock (int x, int y, int width, int height) {
+    public FallingBlock (float x, float y, float width, float height) {
         this.x = x;
         this.y = y;
         this.width = width;
