+++ core/src/com/syurba/combogame/PlacedBlock.java
@@ -3,7 +3,7 @@ package com.syurba.combogame;
 import com.badlogic.gdx.math.Rectangle;
 
 public class PlacedBlock extends Rectangle {
-    public PlacedBlock(int x, int y, int width, int height) {
+    public PlacedBlock(float x, float y, float width, float height) {
         this.x = x;
         this.y = y;
         this.width = width;
